slack: refactored code
