+++ slack/src/data/events.json
