+++ slack/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod global;
+pub mod models;
+pub mod utils;
+
+pub mod routes;
