+++ slack/src/routes.rs
@@ -1,0 +1,8 @@
+pub mod auth;
+pub mod channels;
+pub mod event;
+pub mod inter;
+pub mod join_channel;
+pub mod list_actions;
+pub mod list_events;
+pub mod post;
