+++ slack/src/routes/list_actions.rs
@@ -1,0 +1,7 @@
+use axum::response::IntoResponse;
+
+use crate::global::ACTIONS;
+
+pub async fn list_actions() -> impl IntoResponse {
+    ([("content-type", "application/json")], *ACTIONS)
+}
