+++ slack/src/routes/list_events.rs
@@ -1,0 +1,7 @@
+use axum::response::IntoResponse;
+
+use crate::global::EVENTS;
+
+pub async fn list_events() -> impl IntoResponse {
+    ([("content-type", "application/json")], *EVENTS)
+}
